Fixxed Coords of objects
I realized that objects have coordinates with respect to there parents, so a child's (0, 0, 0) is not the origin of the game, but rather the location of the parent. I fixed many issues caused by this and scaled things to appropriate sizes. I also made the river barrier much nicer and symmetrical.