Health Bars/variety arenas
I added a health bar above a unit. It will be subject to change as we develop the code for certain units, and other things like hiding it until a unit has taken damage, but its functional. Right now if you start the game and hit the spacebar the health bar will take 10 points of damage.

2 Scripts were added into the Units folder, and 1 of the units were given the new components.

A variety of alternate arenas are also to the left and right of our original right now as well.