Added the more less noticeable parts of unit targeting, along with unit rotation and unit death
Added a priority for units. Right now the only priority any game had was structure priority.
Added rotation and a rotation speed for units/towers (this value could be a constant, or a hidden value for every unit. Currently its a constant, but a hidden value change can be made easily)
Added towers, which rotate towards target and reset to center when done
Added unit death
Added basic animation that came with skeletons. More or less just to practice with the idea and to kinda see what the units are doing. It currently does not time with units actually doing damage, as this is more work than its worth for a test unit.
Added a few less notable targeting mechanics. Such mechanics include the idea of an 'attack ready percentage' (ARP) and a 'Maximum attack angle' (MAA).
If a unit is 90% (number subject to change) 'charged' with their attack, they will finish their attack even if its target leaves its range, assuming its still within vision and there is no closer unit.
If a unit is within range of its target, but its angle of attack greater than the MAA, it will only charge its attack up to the ARP until it does rotate enough. (Think of the slow rotating airship)

Some problems faced:
Units have no respect for others and will attempt to walk through them, but will turn if it has to. For normal units, this is good, but for towers and structures, this is not ideal. Trying to figure out a good solution to this.
The scenario where, for example, a flying unit is chasing a ground melee unit. If the flying unit is faster but the ground unit is moving away from the flying unit, the flying unit will enter a constant loop of: moving forward, stopping to charge attack, ground unit leaves range so attack charge resets. Not sure the best solution to this.