Added basic UI for Cards
Basic UI was added, but it does not scale with resolution. I only needed the UI so I can work on its functionality first.

Players now have a resource
Cards work, meaning they place units, take from player resource, and when hovering a card over the arena, it displays a preview of a unit. The player may choose to cancel and they can put the card back over their others.

Some problems I need to solve:
Currently, the unit preview willl go through the tower rather than move around it. The actual placed unit is fine. I know why this is caused, just not sure how to fix it.

Also, some of the UI and the beginning of the game, before a card is played is visible in the scene, but not the view of the camera. I do not know why, but at the moment it doesnt seem like a big deal.