No play zone and preview fix
Added the no play zone to the arena. This no play zone should be future proof, unless we decide that the left zone is not the same size as the right zone.

Also fixed previews from colliding with towers, and this is future proof. Only thing that will need to be updated is if the central tower is thicker.

Added multi unit cards