Abilities Added!
Added skillshot ability and ability UI. Currently, the abilities can detect if it pierces, and if it self destructs/explodes in the ability preview. No aoe damage is implemented yet tho.

Im pushing this now because I need to remove the navmesh agent from the projectile, but this will temporarily break it, so I want to save a working version first.