Boomerang and bugfixes. Working on freeze
Boomerang has been added to the list of properties for a projectile.
Units and towers are ready to be frozen, but nothing currently freezes yet.

Bug fixes:
Fixed targeting errors and attacking errors. There errors included:
    - A unit not doing damage when attacking a target out of range but within the
      chare limit.
    - A unit not targeting correctly when nothing was in range but within vision
Main keep tower health bar will now be inactive at the start of the game
Self destruct not working properly when pierce was on or off, dont remember which