Summon ability has been added, Other fixes below
A new type of ability has been added, with the creative name of 'Create At Location'.
This ability creates a projectile at the targeted location.
Another use of this type is the summon.

With the addition of the summon, buildings now create obstacles that units must walk around, but will not be aloud to create total blockages.

There are two kinds of buildings, small and large.
These buildings cant create blockages due to the new navMeshes added just for buildings. This navMesh is attached to the ground object.

Rare bug fix: There was a small chance that the inRange stat doesnt get updated due to tiny delays in updates. This has been fixed with the addition of the inRangeTargets attribute.