Reworked Card script, now allows drop skills
The card script and the spell script has been merged to a single script, which will now allows for a card to have a unit and multiple abilities. This makes it possible to have drop skills on units. We will likely need 'untargetable' and 'undamagable' timers for units and abilities. The drop skill cannot be stopped regardless if this is done or not.

Cards are also now given a placement restriction stat, which will allow some cards to bypass the spawn zone restriction