EverDawn_Tateki_1
\<Beginning of development under name: "EverDawn">

This changelog will largely be dealing with fixing issues that occurred due to the branch merge.

[Start of Changes]

Moved `Menu Camera` up to Y = 200 just to get it out of the way of the gameplay scene. (and out of the river).

Created a new Scene: `Title_Scene`

The title screen functions as a typical title screen but will likely also have other features that may be needed in case of a damaged download or something of the like. (Creating some sort of method to clear cache for example if there happens to be an issue that prevents the game from booting into the main menu for whatever reason. I've seen it happen in some of the mobile games I play and I'd rather play it safe than sorry to get features like this implemented in case of an issue.) The scene is then intended to be unloaded after the main menu is loaded and other functionality is confirmed to have loaded correctly.

Added an end game/surrender button into `InGameScene` this just unloads `InGameScene` by directly loading `Main_Menu_Scene`

Created `End_Game_Canvas` in `InGameScene` to eventually display game results and etc. It should be activated after the game ends. It also has a button to load back into the main menu.

Created `EndGame()` in `GameManager.cs` which runs when a player's score reaches 3. It destroys all GameObjects and TowerObjects that the game manager has in its respective lists and removes them from said lists. This is not yet implemented with the timer or if the player's keep dies.

It should then hide the gameplay HUD and display the end of game screen. but man am I having some trouble with object references so this part of the code is currently commented out

Added text to the game queue/loading overlay

Added Pot's meme logo into the project as the title screen as `Title_Screen_Background_Moon` which has an attached script `Moon_Background_Temp_Animator` which basically just makes it act like a panorama.

Preemptively changed the `Scene_Loader_Unloader` script to make it a singleton. This is commented out as other objects that need it don't have the functionality to use it as a singleton yet though. Eventually, this object should be used as a singleton that's instantiated at boot or something, opposed to existing in all scenes independently. It's also a prefab now that is added in all scenes under the "loading" type of empty.

Created a basic singleton type script called `Data_Storage` that is capable of serializing a file through C#'s built in means. This was more meant as a test to see how C# goes about serializing files and uses a super basic like cookie cutter type method of going about it. All it currently does is create a binary file which stores a list which could theoretically be used to store what cards are selected. There is not currently anything implemented that uses this script.

The bottom menu selection buttons are no longer setup as a toggle group, but are now managed by the `UI_Tab_Group` and `UI_Tab_Button` scripts. These scripts should allow for flexible Tab System implementation. I'm to expand the functionality of these scripts in the future (ex: probably also going to add color options eventually).

Added a couple sprites - some jokes, some for the tab system, etc. One is a slightly edited version of one of Noden's concept arts. (Changed to be 16x9 resolution and gave it a transparent background.) I just needed something a little closer to what we'd actually be working with so I can work on adjusting the positioning and stuff of character backgrounds in the future.