Added armor stat and limited many variables
Added an armor stat to units.

Many variables have been given a range or a minimum variable, mostly 0.

Minor bug fix:
If a summoner died before the summon has any time to be created, the summon is supposed to be destroyed. Before the projectile stayed and did nothing, now its properly gets deleted.