EverDawn-Tateki-2
Changed some scripts to use (usually serialized) private variables when possible.

Changed naming conventions. I have removed `_` in file names, object names, etc and it is now reserved for denoting object types in their names with a prefix. And organized files in the asset folder. (The _'s are staying in the changelog's names tho.)

Removed tags: `UI_Button` and `UI_Text`. I may add these in as layers later.

Got rid of Scene Loading objects in all scenes, deleted the loader/unloader prefab that was in the project. There is now a scriptable object called `SceneLoader` which is capable of handling all previous functionality and just needs to be called from object in the asset folder. All buttons that handle scene loading have been swapped over to this scheme.

- Currently all buttons are not additively loading scenes - this will likely need to be changed once server infrastructure is put into place, as there will likely be a boot/initialization scene that must remain loaded at all times.
  - (As a personal aside I do think it would be neat if we could set things up where there is an offline mode or something and only the game could still boot and do some things without a network connection. I'll also put this somewhere other than a changelog since I don't know if anyone is actually reading these.)

Moved the `LeanTween` asset folder to the root asset folder instead of `MenuAssets`.

Began working on character implementation to allow for deck building via scriptable objects.

- Currently the SO structure uses `SO_Character` as a sort of "base" type, building cards off that, and then units off of that. (Character > Card > Unit). This can and likely will be broken down further as more scriptable objects are implemented
- May combine character and card databases into a single object, but I've opted to keep them separate for now.

Implemented system in the Units Menu to allow for the automatic importing and displaying of cards.

- `UnitDisplayViewport` has a child object called `CardDatabaseDisplay` which has the attached monobehavior `CardDatabaseDisplay`. This takes the `SO_CardDatabase` scriptable object and the `UnitCard` prefab and creates cards from the list stored in `SO_CardDatabase`. All the card data itself is stored in the `SO_Card` type scriptable object, which can be changed to update the card's parameters.
  - coupled with the grid layout this is capable of autofilling the list of cards in a nice manner by simply adding them to the database.

Started reworking the `InputAnimator` script. Added toggles to turn both the tap animation and cursor trail off.

Began looking at changing UI text components to TextMeshPro. This did install the `TextMesh Pro` folder in the root asset folder.

​	done with title sceen, not gonna touch ingame

# todo

Appended get/setters to public fields (when necessary).

Change text components in the main menu scene to TMP types

finish implementation of drag handlers in `InputAnimator` to play the cursor trail

need to remove `SO_Deck` as it cannot be directly used to serialize a deck during runtime and should probably be changed to a basic class type

finish implementation of `CardDatabaseDisplay` where it can control click events that occur on the cards (this may or may not be possible, partial implementation is completed, was researching the rest)

finish loading animation with its orbit script which is currently under the persistent canvas