EverDawn_Tateki_3
Text in `MainMenuScene` has been switched over to TMP UI text from basic default unity text. Currently I have it set to auto size so some things are wonky currently, I'm not really going to worry about the sizing for now though.

`CardDatabaseDisplay` is now exactly what the name implies, it only displays the cards. Previously there was some functionality in the script that allowed for using the instantiated cards as buttons. This has been decoupled and is now going to be controlled by a separate script which is tied into the event and/or observer system. As a result, I'm considering `CardDatabaseDisplay` complete as of current.

[WIP] Orbit animation works, but needs further work to ensure that the object resets properly after is is disabled

The event system is mostly implemented - it exists mostly for passing data between objects. Currently the two data types it can transfer are Void (nothing), and int. Adding a new data type does require creating ~3 new scripts for each type, but it's easy to do and supports a that entire data type.

Added `cardstats` field to `SO_Card` definition (was previously unit type as a placeholder) which will likely be used when branches get remerged so milk and I can work out how our different card/unit definitions will work together

[todo] (carry over from EverDawn_Tateki_2)

Appended get/setters to public fields (when necessary).

[low priority] finish implementation of drag handlers in `InputAnimator` to play the cursor trail

need to remove `SO_Deck` as it cannot/should not be directly used to serialize a deck during runtime and should be changed to a different class type