Added the ability to supply a custom path curve
Added a script to supply a custom path, image, and collider.

In order to make the above change work, major changes have been made to the skillshot and target script.
The previews made from these scripts are now a bit different in order to allow for these custom paths

Added the ability for an ability to skip the last projectiles given a miss

Added a stunned effect, the same thing as freeze but doesn't lock their animations or turn units blue