EverDawn_Tateki_4
The event system remains (I believe) unchanged from v3, it works; the 3 supported data types are void, int, and so_card currently.

Background swapping now works with the event system and is saved to player prefs.

[WIP] Deck building *works*, but is wonky. Currently to change a deck the user **must** click the card in the lower part of the units menu, and then upon clicking a card in the above 'deck' it will swap out the *first* card in the list that has a matching ID. There still needs to be checks put into place that prevent the user from selecting multiple of the same card. Cards are currently not removed from the lower section upon adding them to a deck. There also needs to be checks to make sure a deck contains 8 and only 8 cards. Decks get saved to a file determined by `Application.persistentDataPath`. The creation of multiple decks is currently not possible.

I'm likely going to split the deck display script into a second one that is more focused on being a mediator between the lower display and the upper, deck, display. This may be a scriptable object, the deck display object, an otherwise empty game object in the scene, or some combination of the above.

[todo] (carry over from EverDawn_Tateki_3)

[WIP] Orbit animation works, but needs further work to ensure that the object resets properly after is is disabled

Appended get/setters to public fields (when necessary).

[low priority] finish implementation of drag handlers in `InputAnimator` to play the cursor trail

need to remove `SO_Deck` as it cannot/should not be directly used to serialize a deck during runtime and should be changed to a different class type