EverDawn_Tateki_5
fixed the reference resolution on all canvases having swapped x and y sizes

changed all canvases canvas scaler components' `Screen Match Mode` to be `Match Width Or Height` and set `Match` to be `.5`

as a consequence of the above: rebuilt/overhauled most canvases' layouts so that they scale with most aspect ratios. They are also less dependent on horizontal/vertical layout groups now. Layouts are largely the same, but have somewhat different sizes from previous versions

changed the auto-size max/min sizes of many text components to accommodate the changes; all have a min text value of 18 now (subject to change)

---

fixed bug where last card in the deck could not be changed

bounded the (more or less) infinitely scrollable scroll rects in the swap background character, and units menu - this still needs a better solution as I had to manually set all the object sizes just based on how many cards we have

added a team preview to the play game overlay

changed name of the project

added a toggle to DeckDisplay to change whether or not a deck is saved when the object is disabled

---

todo:

improve deck building thingy

add an average resource counter to the team formation screen

card/unit details

rework how menu changing works (again)

---

[todo] (carry over from EverDawn_Tateki_3)

[WIP] Orbit animation works, but needs further work to ensure that the object resets properly after is is disabled

Appended get/setters to public fields (when necessary).

[low priority] finish implementation of drag handlers in `InputAnimator` to play the cursor trail

need to remove `SO_Deck` as it cannot/should not be directly used to serialize a deck during runtime and should be changed to a different class type