EverDawn_Tateki_8
renamed MainMenu things to HomeMenu for the sake of consistency in regard to the button name and the scene file's name

Introduced a State Machine system. This is a variable system that utilizes scriptable objects to handle holding states and state changes. There are base abstract classes that can be derived from to create instantiable states and state controllers in ordered to control multiple systems using the same basic logic.

Menu changing has been transitioned to utilize the new state machine architecture. This is more robust and now uses the built in Unity Button type, instead of a custom made one. This is also utilizing the event system to transfer data. Theoretically this should also give us a way to add in extra menus relatively easily.

Introduced scripts to change deck selection to the state machine system. [Not fully implemented]